pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "asyncio>=3.4.3",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "langgraph-checkpoint>=2.1.0",
    "langsmith>=0.4.5",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "tavily-python>=0.7.9",
    "trafilatura>=2.0.0",
]
